+++ apps/react-admin/src/components/roles/create.tsx
@@ -1,0 +1,12 @@
+import { Create, SimpleForm, TextInput, required } from "react-admin";
+
+export function RoleCreate() {
+  return (
+    <Create>
+      <SimpleForm>
+        <TextInput source="name" validate={[required()]} />
+        <TextInput source="description" multiline />
+      </SimpleForm>
+    </Create>
+  );
+}